requirements.txt
# ===== Core ML =====
torch>=2.1.0
transformers>=4.36.0
datasets>=2.16.0
accelerate>=0.25.0
peft>=0.7.0
trl>=0.7.0
bitsandbytes>=0.41.0

# ===== Unsloth (Stages C & B) =====
unsloth[colab-new]>=2024.1

# ===== Evaluation =====
rouge-score>=0.1.2
nltk>=3.8.0
scikit-learn>=1.3.0

# ===== Data =====
pandas>=2.1.0
pyyaml>=6.0

# ===== AWS =====
boto3>=1.34.0

# ===== Deployment =====
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
httpx>=0.26.0

# ===== Notifications =====
requests>=2.31.0

# ===== Dev / Testing =====
pytest>=7.4.0
moto[s3]>=4.2.0
python-dotenv>=1.0.0

# ===== HuggingFace Hub =====
huggingface-hub>=0.20.0
